+++ Sources/TelegramBot/Categories/String+ExtractBotCommand.swift
@@ -1,7 +1,17 @@
-// Telegram Bot SDK for Swift (unofficial).
-// (c) 2015 - 2016 Andrey Fidrya. MIT license. See LICENSE for more information.
+//
+// String+ExtractBotCommand.swift
+//
+// This source file is part of the Telegram Bot SDK for Swift (unofficial).
+//
+// Copyright (c) 2015 - 2016 Andrey Fidrya and the project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See LICENSE.txt for license information
+// See AUTHORS.txt for the list of the project authors
+//
 
 import Foundation
+import ScannerUtils
 
 extension String {
 	/// - Parameter botName: bot name to remove.
@@ -12,14 +22,14 @@ extension String {
         scanner.charactersToBeSkipped = nil
         
         let whitespaceAndNewline = CharacterSet.whitespacesAndNewlines
-        scanner.skipCharactersFromSet(whitespaceAndNewline)
+        scanner.skipCharacters(from: whitespaceAndNewline)
         
         guard scanner.skipString("/") else {
             return self
         }
         
         let alphanumericCharacters = CharacterSet.alphanumerics
-        guard scanner.skipCharactersFromSet(alphanumericCharacters) else {
+        guard scanner.skipCharacters(from: alphanumericCharacters) else {
             return self
         }
 
@@ -35,7 +45,7 @@ extension String {
             "abcdefghijklmnopqrstuvwxyz" +
             "ABCDEFGHIJKLMNOPQRSTUVWXYZ" +
             "1234567890_")
-        guard let username = scanner.scanCharactersFromSet(usernameCharacters) else {
+        guard let username = scanner.scanCharacters(from: usernameCharacters) else {
             // Empty bot name. Treat as no bot name and process the comamnd.
             return self
         }
@@ -45,7 +55,7 @@ extension String {
             return nil
         }
         
-        let t = self as NSString
+        let t = NSString(string: self)
 		return t.substring(to: usernameSeparatorIndex) +
 			t.substring(from: scanner.scanLocation)
     }