+++ Sources/TelegramBot/Protocols/ReplyMarkup.swift
@@ -1,5 +1,14 @@
-// Telegram Bot SDK for Swift (unofficial).
-// (c) 2015 - 2016 Andrey Fidrya. MIT license. See LICENSE for more information.
+//
+// ReplyMarkup.swift
+//
+// This source file is part of the Telegram Bot SDK for Swift (unofficial).
+//
+// Copyright (c) 2015 - 2016 Andrey Fidrya and the project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See LICENSE.txt for license information
+// See AUTHORS.txt for the list of the project authors
+//
 
 import Foundation
 
@@ -12,7 +21,7 @@ extension InlineKeyboardMarkup: ReplyMarkup {
 extension ReplyKeyboardMarkup: ReplyMarkup {
 }
 
-extension ReplyKeyboardHide: ReplyMarkup {
+extension ReplyKeyboardRemove: ReplyMarkup {
 }
 
 extension ForceReply: ReplyMarkup {