+++ Sources/TelegramBot/Router/Arguments.swift
@@ -1,5 +1,14 @@
-// Telegram Bot SDK for Swift (unofficial).
-// (c) 2015 - 2016 Andrey Fidrya. MIT license. See LICENSE for more information.
+//
+// Arguments.swift
+//
+// This source file is part of the Telegram Bot SDK for Swift (unofficial).
+//
+// Copyright (c) 2015 - 2016 Andrey Fidrya and the project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See LICENSE.txt for license information
+// See AUTHORS.txt for the list of the project authors
+//
 
 import Foundation
 
@@ -19,7 +28,7 @@ public class Arguments {
 	}
 	
 	public func scanWord() -> String? {
-		return scanner.scanUpToCharactersFromSet(T.whitespaceAndNewline)
+        return scanner.scanUpToCharacters(from: T.whitespaceAndNewline)
 	}
 	
 	public func scanWords() -> [String] {
@@ -30,13 +39,13 @@ public class Arguments {
 		return words
 	}
 	
-	public func scanInt() -> Int? {
+	public func scanInteger() -> Int? {
 		guard let word = scanWord() else {
 			return nil
 		}
 		let validator = Scanner(string: word)
 		validator.charactersToBeSkipped = nil
-		guard let value = validator.scanInt(), validator.isAtEnd else {
+		guard let value = validator.scanInteger(), validator.isAtEnd else {
 			return nil
 		}
 		return value
@@ -67,13 +76,13 @@ public class Arguments {
 	}
 	
 	public func scanRestOfString() -> String {
-		guard let restOfString = scanner.scanUpToString("") else {
+		guard let restOfString = scanner.scanUpTo("") else {
 			return ""
 		}
 		return restOfString
 	}
 	
 	public func skipRestOfString() {
-		scanner.skipUpToString("")
+		scanner.skipUpTo("")
 	}
 }