+++ Sources/TelegramBot/Router/Context.swift
@@ -1,7 +1,17 @@
-// Telegram Bot SDK for Swift (unofficial).
-// (c) 2015 - 2016 Andrey Fidrya. MIT license. See LICENSE for more information.
+//
+// Context.swift
+//
+// This source file is part of the Telegram Bot SDK for Swift (unofficial).
+//
+// Copyright (c) 2015 - 2016 Andrey Fidrya and the project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See LICENSE.txt for license information
+// See AUTHORS.txt for the list of the project authors
+//
 
 import Foundation
+import Dispatch
 
 public class Context {
 	typealias T = Context