+++ Sources/TelegramBot/TelegramBot.swift
@@ -1,13 +1,23 @@
-// Telegram Bot SDK for Swift (unofficial).
-// (c) 2015 - 2016 Andrey Fidrya. MIT license. See LICENSE for more information.
+//
+// TelegramBot.swift
+//
+// This source file is part of the Telegram Bot SDK for Swift (unofficial).
+//
+// Copyright (c) 2015 - 2016 Andrey Fidrya and the project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See LICENSE.txt for license information
+// See AUTHORS.txt for the list of the project authors
+//
 
 import Foundation
+import Dispatch
 import SwiftyJSON
 
 public class TelegramBot {
     /// `errorHandler`'s completion block type
     /// - SeeAlso: `public var errorHandler: ErrorHandler?`
-    public typealias ErrorHandler = (URLSessionDataTask, DataTaskError) -> ()
+    public typealias ErrorHandler = (URLSessionDataTask, TaskAssociatedData, DataTaskError) -> ()
     
     public typealias DataTaskCompletion = (_ json: JSON, _ error: DataTaskError?)->()
 
@@ -90,7 +100,7 @@ public class TelegramBot {
         NSURLErrorBackgroundSessionWasDisconnected,
     ]
     
-    /// Session. By default, configured with ephemeralSessionConfiguration().
+    /// Session. By default, configured with ephemeral session configuration on OS X or default session configuration on Linux.
     public var session: URLSession
 
     /// Offset for long polling.
@@ -121,6 +131,9 @@ public class TelegramBot {
     /// `defaultErrorHandler` is used by default.
     public var errorHandler: ErrorHandler?
     
+    /// Logging function. Defaults to `print`.
+    public var logger: (_ text: String) -> () = { print($0) }
+    
     /// Defines reconnect delay in seconds depending on `retryCount`. Can be overridden.
     ///
     /// Used by default `errorHandler` implementation.
@@ -149,16 +162,12 @@ public class TelegramBot {
     /// Implements the default error handling logic. Consult
     /// `TelegramBot` class description for details.
     public lazy var defaultErrorHandler: ErrorHandler = {
-        [weak self] task, error in
+        [weak self] task, taskAssociatedData, error in
         
         guard let originalRequest = task.originalRequest else {
             fatalError("\(error)")
         }
  
-        guard let taskAssociatedData = task.associatedData else {
-            fatalError("\(error)")
-        }
-        
         // Strong capture the self if it's still not nil
         guard let actualSelf = self else { return }
 
@@ -172,10 +181,10 @@ public class TelegramBot {
         case let .genericError(_, _, networkError as NSError)
             where networkError.domain == NSURLErrorDomain &&
                 TelegramBot.autoReconnectCodes.contains(networkError.code):
-            print("Network error: \(networkError.localizedDescription)")
+            actualSelf.logger("Network error: \(networkError.localizedDescription)")
             break
         case let .invalidStatusCode(statusCode, _, _) where statusCode != 401: // == 502
-            print("Error: \(error.debugDescription)")
+            actualSelf.logger("Error: \(error.debugDescription)")
             break
         default:
             taskAssociatedData.completion?(nil, error)
@@ -191,12 +200,12 @@ public class TelegramBot {
         // but startDataTaskForRequest will start the new
         // dataTask from workQueue.
         if reconnectDelay == 0.0 {
-            print("Reconnect attempt \(taskAssociatedData.retryCount), will retry at once")
+            actualSelf.logger("Reconnect attempt \(taskAssociatedData.retryCount), will retry at once")
             actualSelf.startDataTaskForRequest(originalRequest, associateTaskWithData: taskAssociatedData)
         } else {
-            print("Reconnect attempt \(taskAssociatedData.retryCount), will retry after \(reconnectDelay) sec")
-            // Be aware that dispatch_after does NOT work correctly with serial queues.
-            // The queue will perform async blocks BEFORE those inserted via dispatch_after.
+            actualSelf.logger("Reconnect attempt \(taskAssociatedData.retryCount), will retry after \(reconnectDelay) sec")
+            // Be aware that asyncAfter does NOT work correctly with serial queues.
+            // The queue will perform async blocks BEFORE those inserted via asyncAfter.
             // So, use global queue for the pause, then execute the actual request on workQueue.
             DispatchQueue.global().asyncAfter(
                 deadline: DispatchTime.now() + reconnectDelay) {
@@ -207,7 +216,11 @@ public class TelegramBot {
     
     /// Default handling of network and parse errors.
     public static let defaultSession: URLSession = {
+        #if os(Linux)
+        let configuration = URLSessionConfiguration.default
+        #else
         let configuration = URLSessionConfiguration.ephemeral
+        #endif
         return URLSession(configuration: configuration)
     }()
     
@@ -285,16 +298,19 @@ public class TelegramBot {
     public func startDataTaskForEndpoint(_ endpoint: String, parameters: [String: Any?], completion: @escaping DataTaskCompletion) {
         let endpointUrl = urlForEndpoint(endpoint)
         let data = HTTPUtils.formUrlencode(parameters)
-		print("endpoint: \(endpoint), data: \(data)")
+		logger("endpoint: \(endpoint), data: \(data)")
         
-        let request = NSMutableURLRequest(url: endpointUrl)
+        var request = URLRequest(url: endpointUrl)
         request.cachePolicy = .reloadIgnoringLocalAndRemoteCacheData
         request.httpMethod = "POST"
 		request.httpBody = data.data(using: String.Encoding.utf8)
+        // Temporarily workaround https://bugs.swift.org/browse/SR-2617
+        #if !os(Linux)
         request.setValue("application/x-www-form-urlencoded", forHTTPHeaderField: "Content-Type")
+        #endif
         
         let taskAssociatedData = TaskAssociatedData(completion)
-        startDataTaskForRequest(request as URLRequest, associateTaskWithData: taskAssociatedData)
+        startDataTaskForRequest(request, associateTaskWithData: taskAssociatedData)
     }
     
     /// Use this function for implementing retrying in
@@ -309,38 +325,35 @@ public class TelegramBot {
         self.workQueue.async {
             var task: URLSessionDataTask?
 			task = self.session.dataTask(with: request) { dataOrNil, responseOrNil, errorOrNil in
-				self.urlSessionDataTaskCompletion(task!, dataOrNil, responseOrNil, errorOrNil)
+				self.urlSessionDataTaskCompletion(task!, taskAssociatedData, dataOrNil, responseOrNil, errorOrNil)
             }
-            if let t = task {
-                t.associatedData = taskAssociatedData
-                t.resume()
-            }
+            task?.resume()
         }
     }
 		
-    private func urlSessionDataTaskCompletion(_ task: URLSessionDataTask, _ dataOrNil: Data?, _ responseOrNil: URLResponse?, _ errorOrNil: Error?) {
+    private func urlSessionDataTaskCompletion(_ task: URLSessionDataTask, _ taskAssociatedData: TaskAssociatedData, _ dataOrNil: Data?, _ responseOrNil: URLResponse?, _ errorOrNil: Error?) {
         
         if let error = errorOrNil {
-            errorHandler?(task, .genericError(
+            errorHandler?(task, taskAssociatedData, .genericError(
                 data: dataOrNil, response: responseOrNil, error: error))
             return
         }
         
         guard let response = responseOrNil as? HTTPURLResponse else {
-            errorHandler?(task, .invalidResponseType(
+            errorHandler?(task, taskAssociatedData, .invalidResponseType(
                 data: dataOrNil, response: responseOrNil))
             return
         }
         
         if response.statusCode != 200 {
-            errorHandler?(task, .invalidStatusCode(
+            errorHandler?(task, taskAssociatedData, .invalidStatusCode(
                 statusCode: response.statusCode,
                 data: dataOrNil, response: response))
             return
         }
         
         guard let data = dataOrNil else {
-            errorHandler?(task, .noDataReceived(
+            errorHandler?(task, taskAssociatedData, .noDataReceived(
                 response: response))
             return
         }
@@ -354,7 +367,7 @@ public class TelegramBot {
         }*/
         
         if !telegramResponse.ok {
-            errorHandler?(task, .serverError(
+            errorHandler?(task, taskAssociatedData, .serverError(
                 telegramResponse: telegramResponse, data: data, response: response))
             return
         }
@@ -369,16 +382,12 @@ public class TelegramBot {
         // task, call it. Completion handler is stored as
         // an associated property and not passed as a function
         // argument to support retrying.
-        if let taskAssociatedData = task.associatedData {
-            
-            // Success
-            if taskAssociatedData.retryCount != 0 {
-                taskAssociatedData.retryCount = 0
-                print("Reconnected to Telegram server")
-            }
-
-            taskAssociatedData.completion?(result, nil)
+        if taskAssociatedData.retryCount != 0 {
+            taskAssociatedData.retryCount = 0
+            logger("Reconnected to Telegram server")
         }
+
+        taskAssociatedData.completion?(result, nil)
     }
 
     private func urlForEndpoint(_ endpoint: String) -> URL {