+++ Sources/TelegramBot/Types/InlineQueryResultCachedMpeg4Gif.swift
@@ -42,7 +42,7 @@ public struct InlineQueryResultCachedMpeg4Gif: JsonConvertible {
         set { json["caption"].string = newValue }
     }
 
-    /// Optional. An Inline keyboard attached to the message
+    /// Optional. Inline keyboard attached to the message
     public var reply_markup: InlineKeyboardMarkup? {
         get {
             let value = json["reply_markup"]