+++ Tests/LinuxMain.swift
@@ -1,0 +1,10 @@
+import XCTest
+@testable import TelegramBotTests
+
+XCTMain([
+     testCase(BlockingServerTests.allTests),
+     testCase(RequestTests.allTests),
+     testCase(RouterTests.allTests),
+     testCase(TelegramBotTests.allTests),
+     testCase(UrlencodeTests.allTests),
+])